working apart from celery
